+++ test/test_system_batch_message.rb
@@ -42,6 +42,8 @@ class TestSystemBatchMessage < Minitest::Test
       end
     end
 
+    assert_no_command_errors(messages)
+
     # Should receive Message::System::Batch, not array
     batch_msg = messages.find { |m| m.is_a?(Rooibos::Message::System::Batch) }
     refute_nil batch_msg, "Should receive System::Batch message"